Check if events exist before deleting eventType

+++ workflow/WorkflowService.go
@@ -422,6 +422,10 @@ func (service *WorkflowService) DeleteEventType(id piazza.Ident) *piazza.JsonRes
 	if hits > 0 || len(triggers) > 0 {
 		return service.statusForbidden(errors.New("Deleting eventTypes that are in use is prohibited"))
 	}
+	events, hits, err := service.eventDB.GetEventsByEventTypeId(id)
+	if hits > 0 || len(events) > 0 {
+		return service.statusForbidden(errors.New("Deleting eventTypes that are in use is prohibited"))
+	}
 
 	ok, err := service.eventTypeDB.DeleteByID(piazza.Ident(id))
 	if !ok {
